# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Availo Ventry - Smart Visitor Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{orgId}/branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List branches of an organization
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create new branch
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchRequest"
      responses:
        "201":
          description: Branch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /organizations/{orgId}/branches/{branchId}:
    get:
      operationId: getBranch
      tags: [branches]
      summary: Get branch by ID
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
        - $ref: "#/components/parameters/branchId"
      responses:
        "200":
          description: Branch details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    put:
      operationId: updateBranch
      tags: [branches]
      summary: Update branch
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
        - $ref: "#/components/parameters/branchId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBranchRequest"
      responses:
        "200":
          description: Branch updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deactivateBranch
      tags: [branches]
      summary: Deactivate a branch
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
        - $ref: "#/components/parameters/branchId"
      responses:
        "200":
          description: Branch deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── USERS ───────────────────────────────────────────────────────────────────
...
  /organizations/{orgId}/invitations:
    get:
      operationId: listInvitations
      tags: [invitations]
      summary: List invitations for an organization
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, expired, revoked]
      responses:
        "200":
          description: List of invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"
    post:
      operationId: createInvitation
      tags: [invitations]
      summary: Invite a new user
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
      responses:
        "201":
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
  /organizations/{orgId}/invitations/{invitationId}/resend:
    post:
      operationId: resendInvitation
      tags: [invitations]
      summary: Resend an invitation
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
        - $ref: "#/components/parameters/invitationId"
      responses:
        "200":
          description: Invitation resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
  /organizations/{orgId}/invitations/{invitationId}/revoke:
    patch:
      operationId: revokeInvitation
      tags: [invitations]
      summary: Revoke a pending invitation
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
        - $ref: "#/components/parameters/invitationId"
      responses:
        "200":
          description: Invitation revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── VISITORS ────────────────────────────────────────────────────────────────
...
  /dashboard/super-admin:
    get:
      operationId: getSuperAdminDashboard
      tags: [dashboard]
      summary: Super Admin platform-wide dashboard
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Super Admin dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuperAdminDashboard"
  /dashboard/org/{orgId}:
    get:
      operationId: getOrgDashboard
      tags: [dashboard]
      summary: Org Admin dashboard
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
      responses:
        "200":
          description: Org dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgDashboard"
  /dashboard/branch/{branchId}:
    get:
      operationId: getBranchDashboard
      tags: [dashboard]
      summary: Branch-level dashboard (Visitor Manager / Receptionist)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/branchId"
      responses:
        "200":
          description: Branch dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BranchDashboard"
  /dashboard/host:
    get:
      operationId: getHostDashboard
      tags: [dashboard]
      summary: Host Employee dashboard
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Host dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostDashboard"

  # ─── AUDIT LOGS ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    orgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
    branchId:
      name: branchId
      in: path
      required: true
      schema:
        type: string
...
    invitationId:
      name: invitationId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    UserRole:
      type: string
      enum:
        - super_admin
        - org_admin
        - visitor_manager
        - receptionist
        - host_employee
    OrgStatus:
      type: string
      enum:
        - active
        - suspended
        - pending_setup
        - deactivated
    OrgType:
      type: string
      enum:
        - government
        - enterprise
        - smb
    SubscriptionTier:
      type: string
      enum:
        - starter
        - professional
        - enterprise
        - government
    EntryMode:
      type: string
      enum:
        - staffed
        - unmanned
        - hybrid
    VerificationPolicy:
      type: string
      enum:
        - none
        - otp_only
        - nafath_only
        - otp_or_nafath
        - nafath_required_otp_fallback
    VisitRequestStatus:
      type: string
      enum:
        - pending
        - approved
        - rejected
        - checked_in
        - checked_out
        - expired
        - cancelled
    VisitRequestType:
      type: string
      enum:
        - pre_registered
        - walk_in

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        logo:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        type:
          $ref: "#/components/schemas/OrgType"
        subscriptionTier:
          $ref: "#/components/schemas/SubscriptionTier"
        publicBookingSlug:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/OrgStatus"
        verificationPolicy:
          $ref: "#/components/schemas/VerificationPolicy"
        nafathEnabled:
          type: boolean
        otpEnabled:
          type: boolean
        verificationBypassForTrusted:
          type: boolean
        maxUsers:
          type: integer
        maxBranches:
          type: integer
        telegramChatId:
          type: string
          nullable: true
        primaryContactName:
          type: string
          nullable: true
        primaryContactEmail:
          type: string
          nullable: true
        primaryContactPhone:
          type: string
          nullable: true
        contractStartDate:
          type: string
          format: date
          nullable: true
        contractEndDate:
          type: string
          format: date
          nullable: true
        suspensionReason:
          type: string
          nullable: true
        setupWizardCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - subscriptionTier
        - status
        - verificationPolicy
        - nafathEnabled
        - otpEnabled
        - verificationBypassForTrusted
        - maxUsers
        - maxBranches
        - setupWizardCompleted
        - createdAt
        - updatedAt
...
    Branch:
      type: object
      properties:
        id:
          type: string
        orgId:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        branchCode:
          type: string
        entryMode:
          $ref: "#/components/schemas/EntryMode"
        verificationPolicyOverride:
          $ref: "#/components/schemas/VerificationPolicy"
          nullable: true
        isActive:
          type: boolean
        maxConcurrentVisitors:
          type: integer
          nullable: true
        workingHoursOverride:
          type: object
          nullable: true
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orgId
        - name
        - branchCode
        - entryMode
        - isActive
        - createdAt
    CreateBranchRequest:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        branchCode:
          type: string
        entryMode:
          $ref: "#/components/schemas/EntryMode"
        maxConcurrentVisitors:
          type: integer
          nullable: true
      required:
        - name
        - branchCode
        - entryMode
    UpdateBranchRequest:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        entryMode:
          $ref: "#/components/schemas/EntryMode"
        verificationPolicyOverride:
          $ref: "#/components/schemas/VerificationPolicy"
          nullable: true
        maxConcurrentVisitors:
          type: integer
          nullable: true
        isActive:
          type: boolean

    # ── Users ─────────────────────────────────────────────────────────────────
...
    Invitation:
      type: object
      properties:
        id:
          type: string
        orgId:
          type: string
        branchId:
          type: string
          nullable: true
        invitedById:
          type: string
        email:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        department:
          type: string
          nullable: true
        jobTitle:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, accepted, expired, revoked]
        tokenExpiresAt:
          type: string
          format: date-time
        sentAt:
          type: string
          format: date-time
        acceptedAt:
          type: string
          format: date-time
          nullable: true
        resendCount:
          type: integer
      required:
        - id
        - orgId
        - invitedById
        - email
        - name
        - role
        - status
        - tokenExpiresAt
        - sentAt
        - resendCount
    CreateInvitationRequest:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        branchId:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        jobTitle:
          type: string
          nullable: true
      required:
        - name
        - email
        - role

    # ── Visitors ──────────────────────────────────────────────────────────────
    Visitor:
      type: object
      properties:
        id:
          type: string
        nationalIdNumber:
          type: string
          nullable: true
        fullName:
          type: string
        fullNameAr:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        isBlacklisted:
          type: boolean
        verificationStatus:
          type: string
          enum:
            - pending
            - verified_otp
            - verified_nafath
            - verified_both
            - skipped_trusted
            - unverified
        lastVerifiedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - isBlacklisted
        - verificationStatus
        - createdAt
...
    VisitRequest:
      type: object
      properties:
        id:
          type: string
        orgId:
          type: string
        branchId:
          type: string
        visitorId:
          type: string
        hostUserId:
          type: string
          nullable: true
        purpose:
          type: string
        purposeAr:
          type: string
          nullable: true
        type:
          $ref: "#/components/schemas/VisitRequestType"
        status:
          $ref: "#/components/schemas/VisitRequestStatus"
        scheduledDate:
          type: string
          format: date
        scheduledTimeFrom:
          type: string
          nullable: true
        scheduledTimeTo:
          type: string
          nullable: true
        qrExpiresAt:
          type: string
          format: date-time
          nullable: true
        checkInTime:
          type: string
          format: date-time
          nullable: true
        checkOutTime:
          type: string
          format: date-time
          nullable: true
        approvalMethod:
          type: string
          enum: [web, telegram]
          nullable: true
        approvedById:
          type: string
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        visitor:
          $ref: "#/components/schemas/Visitor"
          nullable: true
        hostUser:
          $ref: "#/components/schemas/UserSummary"
          nullable: true
        branch:
          $ref: "#/components/schemas/BranchSummary"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orgId
        - branchId
        - visitorId
        - purpose
        - type
        - status
        - scheduledDate
        - createdAt
...
    SuperAdminDashboard:
      type: object
      properties:
        totalOrgs:
          type: integer
        activeOrgs:
          type: integer
        suspendedOrgs:
          type: integer
        pendingSetupOrgs:
          type: integer
        totalVisitorsToday:
          type: integer
        totalRequestsToday:
          type: integer
        orgsByTier:
          type: array
          items:
            type: object
            properties:
              tier:
                type: string
              count:
                type: integer
            required:
              - tier
              - count
        recentOrgs:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
        monthlyGrowth:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
              newOrgs:
                type: integer
              totalVisitors:
                type: integer
            required:
              - month
              - newOrgs
              - totalVisitors
      required:
        - totalOrgs
        - activeOrgs
        - suspendedOrgs
        - pendingSetupOrgs
        - totalVisitorsToday
        - totalRequestsToday
        - orgsByTier
        - recentOrgs
        - monthlyGrowth
    OrgDashboard:
      type: object
      properties:
        visitorsToday:
          type: integer
        pendingApprovals:
          type: integer
        checkedInNow:
          type: integer
        weeklyTrend:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              count:
                type: integer
            required:
              - date
              - count
        branchComparison:
          type: array
          items:
            type: object
            properties:
              branchName:
                type: string
              visitorsToday:
                type: integer
              pendingApprovals:
                type: integer
            required:
              - branchName
              - visitorsToday
              - pendingApprovals
        recentRequests:
          type: array
          items:
            $ref: "#/components/schemas/VisitRequest"
      required:
        - visitorsToday
        - pendingApprovals
        - checkedInNow
        - weeklyTrend
        - branchComparison
        - recentRequests
    BranchDashboard:
      type: object
      properties:
        expectedVisitorsToday:
          type: array
          items:
            $ref: "#/components/schemas/VisitRequest"
        pendingRequests:
          type: integer
        checkedInNow:
          type: integer
        overstayAlerts:
          type: array
          items:
            $ref: "#/components/schemas/VisitRequest"
      required:
        - expectedVisitorsToday
        - pendingRequests
        - checkedInNow
        - overstayAlerts
    HostDashboard:
      type: object
      properties:
        upcomingVisitors:
          type: array
          items:
            $ref: "#/components/schemas/VisitRequest"
        pastVisitors:
          type: array
          items:
            $ref: "#/components/schemas/VisitRequest"
        totalVisitorsThisMonth:
          type: integer
      required:
        - upcomingVisitors
        - pastVisitors
        - totalVisitorsThisMonth

    # ── Audit Logs ────────────────────────────────────────────────────────────
...